Refactored to separate model and results
model.run() now returns a results object. Which contains all parameter information and output of that run, independent of what happens to the model object later.

Removes the need for observing changes to parameter dataclasses, or checking if the model has run or is initialized.

Also updated tests to work with this new behaviour.